# Using FrozenDictionaries in IEndpointAddressScheme implementations