# Using FrozenDictionaries in IEndpointAddressScheme implementations

Using `FrozenDictionary` in types where values of lookup tables don't mutate: `RouteValuesAddressScheme` and `EndpointNameAddressScheme`.

## Description

These types previously mentioned in discussion of the linked issue #47569. In this PR I apply the corresponding change to freeze the dictionary of the cached values. The key comparers match the casing of the original collection.

I use the existing performance test of `SingleRouteRouteValuesAddressSchemeBenchmark` to run before after comparison. The impacted test is case is the *RouteName*.


Before:

```
|      Method |       Mean |     Error |    StdDev |          Op/s | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |-----------:|----------:|----------:|--------------:|------:|--------:|-------:|------:|------:|----------:|
|    Baseline |   4.806 ns | 0.0475 ns | 0.0422 ns | 208,088,445.4 |  1.00 |    0.00 | 0.0003 |     - |     - |      32 B |
| RouteValues | 104.050 ns | 0.9653 ns | 0.9030 ns |   9,610,791.9 | 21.67 |    0.23 | 0.0032 |     - |     - |     304 B |
|   RouteName |  23.047 ns | 0.3551 ns | 0.3322 ns |  43,389,678.3 |  4.80 |    0.07 | 0.0009 |     - |     - |      80 B |
```

After:

```
|      Method |       Mean |     Error |    StdDev |          Op/s | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |-----------:|----------:|----------:|--------------:|------:|--------:|-------:|------:|------:|----------:|
|    Baseline |   4.832 ns | 0.0432 ns | 0.0404 ns | 206,932,638.2 |  1.00 |    0.00 | 0.0003 |     - |     - |      32 B |
| RouteValues | 105.344 ns | 0.8564 ns | 0.7151 ns |   9,492,689.7 | 21.82 |    0.20 | 0.0032 |     - |     - |     304 B |
|   RouteName |  16.333 ns | 0.1244 ns | 0.1164 ns |  61,224,881.1 |  3.38 |    0.03 | 0.0009 |     - |     - |      80 B |
```

Linked: #47569