@ladeak 2023-12-20
cc. @javiercn @JamesNK @BrennanConroy this PR falls into the same space as the previous ones. Here I only ran the existing performance tests that cover the singe route value address scheme. Should I add further benchmarks for other cases, or does this existing benchmark give enough confidence in the perf benefits?