@JamesNK 2023-12-30
Looks good to me. The dictionaries last the lifetime of the endpoint data source, and it should either last the lifetime of the app or change rarely. And the dictionaries are lazily initialized so the one-off startup cost and constant memory cost are only hit if you use link generation.